+++ test/unit/mysql_clone/chk_rep_cfg.py
@@ -22,7 +22,6 @@ import mock
 # Local
 sys.path.append(os.getcwd())
 import mysql_clone                              # pylint:disable=E0401,C0413
-import lib.gen_libs as gen_libs             # pylint:disable=E0401,C0413,R0402
 import version                                  # pylint:disable=E0401,C0413
 
 __version__ = version.__version__
@@ -127,19 +126,32 @@ class UnitTest(unittest.TestCase):
 
     Methods:
         setUp
-        test_config_fail3
-        test_config_fail2
+        test_config_fail_post_8026a
+        test_config_fail_post_8026
+        test_config_fail_pre_8026a
+        test_config_fail_pre_8026
+        test_config_faila
         test_config_fail
-        test_clone_gtid_off3
-        test_clone_gtid_off2
+        test_clone_gtid_off_post_8026a
+        test_clone_gtid_off_post_8026
+        test_clone_gtid_off_pre_8026a
+        test_clone_gtid_off_pre_8026
+        test_clone_gtid_offa
         test_clone_gtid_off
-        test_config_pass3
-        test_config_pass2
+        test_config_pass_post_8026a
+        test_config_pass_post_8026
+        test_config_pass_pre_8026a
+        test_config_pass_pre_8026
+        test_config_passa
         test_config_pass
-        test_mysql_80a
-        test_mysql_80
-        test_no_rep2
-        test_no_rep
+        test_mysql_post_8026a
+        test_mysql_post_8026
+        test_mysql_pre_8026a
+        test_mysql_pre_8026
+        test_no_rep_post_8026a
+        test_no_rep_post_8026
+        test_no_rep_pre_8026a
+        test_no_rep_pre_8026
 
     """
 
@@ -174,7 +186,7 @@ class UnitTest(unittest.TestCase):
         self.results.append("--master-data=2")
         self.results2 = list(self.opt_arg_list)
         self.results2.append("--master-data=1")
-        self.results3 = []
+        self.results3 = list(self.opt_arg_list)
         self.results4 = list(self.opt_arg_list)
         self.results4.append("--source-data=2")
         self.results5 = list(self.opt_arg_list)
@@ -186,9 +198,9 @@ class UnitTest(unittest.TestCase):
     @mock.patch(
         "mysql_clone.mysql_libs.disconnect", mock.Mock(return_value=True))
     @mock.patch("mysql_clone.cfg_chk")
-    def test_config_fail3(self, mock_cfg):
+    def test_config_fail_post_8026a(self, mock_cfg):
 
-        """Function:  test_config_fail3
+        """Function:  test_config_fail_post_8026a
 
         Description:  Test with configuration checks fails post MySQL 8.0.26.
 
@@ -199,18 +211,38 @@ class UnitTest(unittest.TestCase):
         mock_cfg.return_value = False
         self.source.version = self.version3
 
-        with gen_libs.no_std_out():
-            self.assertEqual(
-                mysql_clone.chk_rep_cfg(
-                    self.source, self.clone, self.args, self.req_rep_cfg,
-                    self.opt_arg_list), self.results3)
+        self.assertEqual(
+            mysql_clone.chk_rep_cfg(
+                self.source, self.clone, self.args, self.req_rep_cfg,
+                self.opt_arg_list)[1], False)
 
     @mock.patch(
         "mysql_clone.mysql_libs.disconnect", mock.Mock(return_value=True))
     @mock.patch("mysql_clone.cfg_chk")
-    def test_config_fail2(self, mock_cfg):
+    def test_config_fail_post_8026(self, mock_cfg):
 
-        """Function:  test_config_fail2
+        """Function:  test_config_fail_post_8026
+
+        Description:  Test with configuration checks fails post MySQL 8.0.26.
+
+        Arguments:
+
+        """
+
+        mock_cfg.return_value = False
+        self.source.version = self.version3
+
+        self.assertEqual(
+            mysql_clone.chk_rep_cfg(
+                self.source, self.clone, self.args, self.req_rep_cfg,
+                self.opt_arg_list)[0], self.results3)
+
+    @mock.patch(
+        "mysql_clone.mysql_libs.disconnect", mock.Mock(return_value=True))
+    @mock.patch("mysql_clone.cfg_chk")
+    def test_config_fail_pre_8026a(self, mock_cfg):
+
+        """Function:  test_config_fail_pre_8026a
 
         Description:  Test with configuration checks fails pre MySQL 8.0.26.
 
@@ -221,11 +253,52 @@ class UnitTest(unittest.TestCase):
         mock_cfg.return_value = False
         self.source.version = self.version2
 
-        with gen_libs.no_std_out():
-            self.assertEqual(
-                mysql_clone.chk_rep_cfg(
-                    self.source, self.clone, self.args, self.req_rep_cfg,
-                    self.opt_arg_list), self.results3)
+        self.assertEqual(
+            mysql_clone.chk_rep_cfg(
+                self.source, self.clone, self.args, self.req_rep_cfg,
+                self.opt_arg_list)[1], False)
+
+    @mock.patch(
+        "mysql_clone.mysql_libs.disconnect", mock.Mock(return_value=True))
+    @mock.patch("mysql_clone.cfg_chk")
+    def test_config_fail_pre_8026(self, mock_cfg):
+
+        """Function:  test_config_fail_pre_8026
+
+        Description:  Test with configuration checks fails pre MySQL 8.0.26.
+
+        Arguments:
+
+        """
+
+        mock_cfg.return_value = False
+        self.source.version = self.version2
+
+        self.assertEqual(
+            mysql_clone.chk_rep_cfg(
+                self.source, self.clone, self.args, self.req_rep_cfg,
+                self.opt_arg_list)[0], self.results3)
+
+    @mock.patch(
+        "mysql_clone.mysql_libs.disconnect", mock.Mock(return_value=True))
+    @mock.patch("mysql_clone.cfg_chk")
+    def test_config_faila(self, mock_cfg):
+
+        """Function:  test_config_faila
+
+        Description:  Test with configuration checks fails.
+
+        Arguments:
+
+        """
+
+        mock_cfg.return_value = False
+        self.source.version = self.version
+
+        self.assertEqual(
+            mysql_clone.chk_rep_cfg(
+                self.source, self.clone, self.args, self.req_rep_cfg,
+                self.opt_arg_list)[1], False)
 
     @mock.patch(
         "mysql_clone.mysql_libs.disconnect", mock.Mock(return_value=True))
@@ -243,16 +316,15 @@ class UnitTest(unittest.TestCase):
         mock_cfg.return_value = False
         self.source.version = self.version
 
-        with gen_libs.no_std_out():
-            self.assertEqual(
-                mysql_clone.chk_rep_cfg(
-                    self.source, self.clone, self.args, self.req_rep_cfg,
-                    self.opt_arg_list), self.results3)
+        self.assertEqual(
+            mysql_clone.chk_rep_cfg(
+                self.source, self.clone, self.args, self.req_rep_cfg,
+                self.opt_arg_list)[0], self.results3)
 
     @mock.patch("mysql_clone.cfg_chk")
-    def test_clone_gtid_off3(self, mock_cfg):
+    def test_clone_gtid_off_post_8026a(self, mock_cfg):
 
-        """Function:  test_clone_gtid_off3
+        """Function:  test_clone_gtid_off_post_8026a
 
         Description:  Test with Clone GTID Mode off post MySQL 8.0.26.
 
@@ -268,12 +340,33 @@ class UnitTest(unittest.TestCase):
         self.assertEqual(
             mysql_clone.chk_rep_cfg(
                 self.source, self.clone, self.args, self.req_rep_cfg,
-                self.opt_arg_list), self.results5)
+                self.opt_arg_list)[1], True)
 
     @mock.patch("mysql_clone.cfg_chk")
-    def test_clone_gtid_off2(self, mock_cfg):
+    def test_clone_gtid_off_post_8026(self, mock_cfg):
 
-        """Function:  test_clone_gtid_off2
+        """Function:  test_clone_gtid_off_post_8026
+
+        Description:  Test with Clone GTID Mode off post MySQL 8.0.26.
+
+        Arguments:
+
+        """
+
+        mock_cfg.return_value = True
+        self.source.version = self.version3
+
+        self.clone.gtid_mode = False
+
+        self.assertEqual(
+            mysql_clone.chk_rep_cfg(
+                self.source, self.clone, self.args, self.req_rep_cfg,
+                self.opt_arg_list)[0], self.results5)
+
+    @mock.patch("mysql_clone.cfg_chk")
+    def test_clone_gtid_off_pre_8026a(self, mock_cfg):
+
+        """Function:  test_clone_gtid_off_pre_8026a
 
         Description:  Test with Clone GTID Mode off pre MySQL 8.0.26.
 
@@ -289,7 +382,49 @@ class UnitTest(unittest.TestCase):
         self.assertEqual(
             mysql_clone.chk_rep_cfg(
                 self.source, self.clone, self.args, self.req_rep_cfg,
-                self.opt_arg_list), self.results2)
+                self.opt_arg_list)[1], True)
+
+    @mock.patch("mysql_clone.cfg_chk")
+    def test_clone_gtid_off_pre_8026(self, mock_cfg):
+
+        """Function:  test_clone_gtid_off_pre_8026
+
+        Description:  Test with Clone GTID Mode off pre MySQL 8.0.26.
+
+        Arguments:
+
+        """
+
+        mock_cfg.return_value = True
+        self.source.version = self.version2
+
+        self.clone.gtid_mode = False
+
+        self.assertEqual(
+            mysql_clone.chk_rep_cfg(
+                self.source, self.clone, self.args, self.req_rep_cfg,
+                self.opt_arg_list)[0], self.results2)
+
+    @mock.patch("mysql_clone.cfg_chk")
+    def test_clone_gtid_offa(self, mock_cfg):
+
+        """Function:  test_clone_gtid_offa
+
+        Description:  Test with Clone GTID Mode off.
+
+        Arguments:
+
+        """
+
+        mock_cfg.return_value = True
+        self.source.version = self.version
+
+        self.clone.gtid_mode = False
+
+        self.assertEqual(
+            mysql_clone.chk_rep_cfg(
+                self.source, self.clone, self.args, self.req_rep_cfg,
+                self.opt_arg_list)[1], True)
 
     @mock.patch("mysql_clone.cfg_chk")
     def test_clone_gtid_off(self, mock_cfg):
@@ -310,12 +445,12 @@ class UnitTest(unittest.TestCase):
         self.assertEqual(
             mysql_clone.chk_rep_cfg(
                 self.source, self.clone, self.args, self.req_rep_cfg,
-                self.opt_arg_list), self.results2)
+                self.opt_arg_list)[0], self.results2)
 
     @mock.patch("mysql_clone.cfg_chk")
-    def test_config_pass3(self, mock_cfg):
+    def test_config_pass_post_8026a(self, mock_cfg):
 
-        """Function:  test_config_pass3
+        """Function:  test_config_pass_post_8026a
 
         Description:  Test with configuration checks pass post MySQL 8.0.26.
 
@@ -329,12 +464,31 @@ class UnitTest(unittest.TestCase):
         self.assertEqual(
             mysql_clone.chk_rep_cfg(
                 self.source, self.clone, self.args, self.req_rep_cfg,
-                self.opt_arg_list), self.results4)
+                self.opt_arg_list)[1], True)
 
     @mock.patch("mysql_clone.cfg_chk")
-    def test_config_pass2(self, mock_cfg):
+    def test_config_pass_post_8026(self, mock_cfg):
 
-        """Function:  test_config_pass2
+        """Function:  test_config_pass_post_8026
+
+        Description:  Test with configuration checks pass post MySQL 8.0.26.
+
+        Arguments:
+
+        """
+
+        mock_cfg.return_value = True
+        self.source.version = self.version3
+
+        self.assertEqual(
+            mysql_clone.chk_rep_cfg(
+                self.source, self.clone, self.args, self.req_rep_cfg,
+                self.opt_arg_list)[0], self.results4)
+
+    @mock.patch("mysql_clone.cfg_chk")
+    def test_config_pass_pre_8026a(self, mock_cfg):
+
+        """Function:  test_config_pass_pre_8026a
 
         Description:  Test with configuration checks pass pre MySQL 8.0.26.
 
@@ -348,7 +502,45 @@ class UnitTest(unittest.TestCase):
         self.assertEqual(
             mysql_clone.chk_rep_cfg(
                 self.source, self.clone, self.args, self.req_rep_cfg,
-                self.opt_arg_list), self.results)
+                self.opt_arg_list)[1], True)
+
+    @mock.patch("mysql_clone.cfg_chk")
+    def test_config_pass_pre_8026(self, mock_cfg):
+
+        """Function:  test_config_pass_pre_8026
+
+        Description:  Test with configuration checks pass pre MySQL 8.0.26.
+
+        Arguments:
+
+        """
+
+        mock_cfg.return_value = True
+        self.source.version = self.version2
+
+        self.assertEqual(
+            mysql_clone.chk_rep_cfg(
+                self.source, self.clone, self.args, self.req_rep_cfg,
+                self.opt_arg_list)[0], self.results)
+
+    @mock.patch("mysql_clone.cfg_chk")
+    def test_config_passa(self, mock_cfg):
+
+        """Function:  test_config_passa
+
+        Description:  Test with configuration checks pass.
+
+        Arguments:
+
+        """
+
+        mock_cfg.return_value = True
+        self.source.version = self.version
+
+        self.assertEqual(
+            mysql_clone.chk_rep_cfg(
+                self.source, self.clone, self.args, self.req_rep_cfg,
+                self.opt_arg_list)[1], True)
 
     @mock.patch("mysql_clone.cfg_chk")
     def test_config_pass(self, mock_cfg):
@@ -367,12 +559,12 @@ class UnitTest(unittest.TestCase):
         self.assertEqual(
             mysql_clone.chk_rep_cfg(
                 self.source, self.clone, self.args, self.req_rep_cfg,
-                self.opt_arg_list), self.results)
+                self.opt_arg_list)[0], self.results)
 
     @mock.patch("mysql_clone.cfg_chk")
-    def test_mysql_80a(self, mock_cfg):
+    def test_mysql_post_8026a(self, mock_cfg):
 
-        """Function:  test_mysql_80a
+        """Function:  test_mysql_post_8026a
 
         Description:  Test with MySQL 8.0 version post 8.0.26.
 
@@ -386,12 +578,31 @@ class UnitTest(unittest.TestCase):
         self.assertEqual(
             mysql_clone.chk_rep_cfg(
                 self.source, self.clone, self.args, self.req_rep_cfg,
-                self.opt_arg_list), self.results4)
+                self.opt_arg_list)[1], True)
 
     @mock.patch("mysql_clone.cfg_chk")
-    def test_mysql_80(self, mock_cfg):
+    def test_mysql_post_8026(self, mock_cfg):
 
-        """Function:  test_mysql_80
+        """Function:  test_mysql_post_8026
+
+        Description:  Test with MySQL 8.0 version post 8.0.26.
+
+        Arguments:
+
+        """
+
+        mock_cfg.return_value = True
+        self.source.version = self.version3
+
+        self.assertEqual(
+            mysql_clone.chk_rep_cfg(
+                self.source, self.clone, self.args, self.req_rep_cfg,
+                self.opt_arg_list)[0], self.results4)
+
+    @mock.patch("mysql_clone.cfg_chk")
+    def test_mysql_pre_8026a(self, mock_cfg):
+
+        """Function:  test_mysql_pre_8026a
 
         Description:  Test with MySQL 8.0 version pre 8.0.26.
 
@@ -405,11 +616,30 @@ class UnitTest(unittest.TestCase):
         self.assertEqual(
             mysql_clone.chk_rep_cfg(
                 self.source, self.clone, self.args, self.req_rep_cfg,
-                self.opt_arg_list), self.results)
+                self.opt_arg_list)[1], True)
 
-    def test_no_rep2(self):
+    @mock.patch("mysql_clone.cfg_chk")
+    def test_mysql_pre_8026(self, mock_cfg):
 
-        """Function:  test_no_rep2
+        """Function:  test_mysql_pre_8026
+
+        Description:  Test with MySQL 8.0 version pre 8.0.26.
+
+        Arguments:
+
+        """
+
+        mock_cfg.return_value = True
+        self.source.version = self.version2
+
+        self.assertEqual(
+            mysql_clone.chk_rep_cfg(
+                self.source, self.clone, self.args, self.req_rep_cfg,
+                self.opt_arg_list)[0], self.results)
+
+    def test_no_rep_post_8026a(self):
+
+        """Function:  test_no_rep_post_8026a
 
         Description:  Test with no replication configured post MySQL 8.0.26.
 
@@ -422,11 +652,28 @@ class UnitTest(unittest.TestCase):
         self.assertEqual(
             mysql_clone.chk_rep_cfg(
                 self.source, self.clone, self.args2, self.req_rep_cfg,
-                self.opt_arg_list), self.results4)
+                self.opt_arg_list)[1], True)
 
-    def test_no_rep(self):
+    def test_no_rep_post_8026(self):
 
-        """Function:  test_no_rep
+        """Function:  test_no_rep_post_8026
+
+        Description:  Test with no replication configured post MySQL 8.0.26.
+
+        Arguments:
+
+        """
+
+        self.source.version = self.version3
+
+        self.assertEqual(
+            mysql_clone.chk_rep_cfg(
+                self.source, self.clone, self.args2, self.req_rep_cfg,
+                self.opt_arg_list)[0], self.results4)
+
+    def test_no_rep_pre_8026a(self):
+
+        """Function:  test_no_rep_pre_8026a
 
         Description:  Test with no replication configured pre MySQL 8.0.26.
 
@@ -439,7 +686,24 @@ class UnitTest(unittest.TestCase):
         self.assertEqual(
             mysql_clone.chk_rep_cfg(
                 self.source, self.clone, self.args2, self.req_rep_cfg,
-                self.opt_arg_list), self.results)
+                self.opt_arg_list)[1], True)
+
+    def test_no_rep_pre_8026(self):
+
+        """Function:  test_no_rep_pre_8026
+
+        Description:  Test with no replication configured pre MySQL 8.0.26.
+
+        Arguments:
+
+        """
+
+        self.source.version = self.version2
+
+        self.assertEqual(
+            mysql_clone.chk_rep_cfg(
+                self.source, self.clone, self.args2, self.req_rep_cfg,
+                self.opt_arg_list)[0], self.results)
 
 
 if __name__ == "__main__":